added header descriptions and new i2c api header
