+++ firmware/CAN-DevBoard/CMakeLists.txt
@@ -42,7 +42,7 @@ set(PROJECT_INCLUDE_DIRECTORIES
     ${CMAKE_CURRENT_SOURCE_DIR}
     ${USER_DIR}/inc
     ${RESOURCES_DIR}/Drivers/SG_Drivers/CAN/Include
-    ${RESOURCES_DIR}/Drivers/SG_Drivers/BMS/Include)
+    ${RESOURCES_DIR}/Drivers/SG_Drivers/Include)
 
 # Sources
 file(GLOB STM32CUBEMX_SOURCES
@@ -55,7 +55,7 @@ file(GLOB STM32CUBEMX_SOURCES
 file(GLOB_RECURSE PROJECT_SOURCES FOLLOW_SYMLINKS
     ${USER_DIR}/src/*.cpp
     ${USER_DIR}/src/*.c
-    ${RESOURCES_DIR}/Drivers/SG_Drivers/*/Src/*.cpp)
+    ${RESOURCES_DIR}/Drivers/SG_Drivers/Src/*.cpp)
 
 # Executable files
 add_executable(${EXECUTABLE}
