+++ firmware/NUCLEOL476RG-FreeRTOS-ExampleProject/CMakeLists.txt
@@ -39,7 +39,8 @@ set(CUBEMX_INCLUDE_DIRECTORIES
     ${RESOURCES_DIR}/Middlewares/Third_Party/FreeRTOS/Source/portable/GCC/ARM_CM4F)
 set(PROJECT_INCLUDE_DIRECTORIES
     ${CMAKE_CURRENT_SOURCE_DIR}
-    ${USER_DIR}/inc)
+    ${USER_DIR}/inc
+    ${RESOURCES_DIR}/Drivers/SG_Drivers/Include)
 
 # Sources
 file(GLOB STM32CUBEMX_SOURCES
@@ -52,7 +53,8 @@ file(GLOB STM32CUBEMX_SOURCES
     ${RESOURCES_DIR}/Middlewares/Third_Party/FreeRTOS/Source/portable/GCC/ARM_CM4F/port.c)
 file(GLOB_RECURSE PROJECT_SOURCES FOLLOW_SYMLINKS
     ${USER_DIR}/src/*.cpp
-    ${USER_DIR}/src/*.c)
+    ${USER_DIR}/src/*.c
+    ${RESOURCES_DIR}/Drivers/SG_Drivers/Src/*.cpp)
 
 # Executable files
 add_executable(${EXECUTABLE}
